track id: [https://trac.osgeo.org/fdo/ticket/113]: Code changes related to RFC 5 and 8: This code drop adds the implementation of the functions NullValue, ToDouble, ToFloat, ToInt32 and ToInt64 and modifies some other function implementation with respect to exception handling (in this one exception message is replaced with a different one).
git-svn-id: http://svn.osgeo.org/fdo/trunk@3327 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Utilities/ExpressionEngine/Src/Functions/String/FdoFunctionLtrim.cpp
@@ -153,7 +153,7 @@ FdoLiteralValue *FdoFunctionLtrim::Evaluate (
         return FdoStringValue::Create(base_string);
     else
       return FdoStringValue::Create(
-                                base_string.Mid((size_t) pos, string_length));
+                    base_string.Mid((size_t) pos, (size_t) string_length));
 
 }  //  Evaluate ()
